+++ tests/infra/network.py
@@ -18,7 +18,6 @@ from infra.tx_status import TxStatus
 from ccf.tx_id import TxID
 import random
 from dataclasses import dataclass
-from math import ceil
 import http
 import pprint
 import functools