+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Removed
+
+- Removed deprecated `set_execute_outside_consensus()` API (#3886, #3673).
+
 ### Changed
 
 ### Added