+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - The node-to-node interface configuration now supports a `published_address` to enable networks with nodes running in different (virtual) subnets (#3867).
-- Added a `/node/service/previous_identity` endpoint (#3880).
+- Added a `/node/service/previous_identity` endpoint (#3880), which can be used during a recovery to look up the identity of the service before the catastrophic failure.
 
 ## [2.0.0]